fix norm issues with cone implementation
